refactor: 게시글, 댓글 서비스를 추상화 후 impl 구현하도록 수정
